requirements.txt
Django==4.2.13
djangorestframework==3.15.1
drf-spectacular==0.27.2
gunicorn==20.1.0
psycopg2-binary
python-dotenv==1.0.1
PyYAML==6.0.1
drf-nested-routers==0.94.1